Fix remove HT config from bootstrap and app classloader
