+++ javaagent/build.gradle.kts
@@ -12,7 +12,6 @@ dependencies {
     // https://oss.jfrog.org/artifactory/oss-snapshot-local/io/opentelemetry/instrumentation/auto/
     // https://dl.bintray.com/open-telemetry/maven/
     implementation("io.opentelemetry.javaagent", "opentelemetry-javaagent", version = "${versions["opentelemetry_java_agent"]}", classifier = "all")
-    implementation(project(":javaagent-core"))
     implementation(project(":filter-api"))
 }
 
@@ -50,8 +49,7 @@ tasks {
         }
 
         relocate("org.slf4j", "io.opentelemetry.javaagent.slf4j")
-        // TODO causes data not being reported
-//        relocate("java.util.logging.Logger", "io.opentelemetry.javaagent.bootstrap.PatchLogger")
+        relocate("java.util.logging.Logger", "io.opentelemetry.javaagent.bootstrap.PatchLogger")
 
         // prevents conflict with library instrumentation
         relocate("io.opentelemetry.instrumentation.api", "io.opentelemetry.javaagent.shaded.instrumentation.api")
